PagingListView can now reload if requested
Before, after clearing the list, setting hasMoreItems to true had no
effect on reloading. This removes the item count check and re-adds the
footer view if hasMoreItems is set to true.

This also updates the gradle build (newer android gradle version isn't backwards
compatible and broke the build process).

Updated the sample as well to set a refresh button